py3requirements.txt
Flask
flask-mysqldb
DateTime
Flask-Limiter
Flask-Mail
setuptools
Jinja2
SQLAlchemy
pymysql
Werkzeug
decorator
ipython
itsdangerous
numpy
pandas
pandasql
pyOpenSSL
python-dateutil
pytz
requests
sasl
scipy
setuptools
thrift
urllib3
py3-validate-email
pytemperature
redis[hiredis]
